Change `@shim` annotation to `@missing_method`
My understanding is that the meaning of `@shim` annotation is that the method never really exists at runtime. On the contrary `@missing_method` is for signalling that the method exists but only via `method_missing` magic, so would not appear in `instance_methods`.

In this case, these methods are handled using `method_missing` so the better annotation would be to use `@missing_method` with an explanation of what they are delegated to.