squash! Validate @method_missing on singleton methods
Using `singleton_method` instead of `method` fixes a difference in
behavior between Ruby 2.7 and 3.1.

```rb
class Foo; end
Foo.method(:method_missing)
```

In 2.7 this returns `#<Method: Foo.method_missing(*)>` and in 3.1 it
returns `#<Method: #<Class:Foo>(BasicObject)#method_missing(*)>`. The
latter is correctly rejected when we match on `/BasicObject/` but the
former isn't.

Because we know we're dealing with a singleton method, we can use
`singleton_method` here instead, which raises a `NameError` in both 2.7
and 3.1.