Add `JSON::Ext::ParserConfig` constant to json gem annotation
This constant was added to the json gem's C extension in version 2.10.0
(see https://github.com/ruby/json/commit/c7aca66d8a29941d0ab1b5753a3c69e007d46154),
and Tapioca is not able to find it while generating types for the json gem. By
including it in the gem annotation, we can make it available to anybody using
a newer version of the json gem.