Improve typing of `Stripe::ApiResource#metadata`
Per the docs, values are always strings (there is no nesting). I opted to not change the signature of `#metadata=` because you can technically pass in non-strings to the method and either the values will be stringified (e.g. `1` becomes `"1"`) or the keys ignored (e.g. a key with a `nil` value).

https://docs.stripe.com/metadata